requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
plotly>=5.18.0
pyyaml>=6.0
python-dotenv>=1.0.0
google-ads>=24.0.0
google-auth-oauthlib>=1.0.0
bingads>=13.0.18
requests>=2.31.0
